Update to gldf 1.0.0-rc.1

+++ src/Gldf.Net/Domain/Definition/Types/Rotation.cs
@@ -2,12 +2,12 @@
 {
     public class Rotation
     {
-        public int X { get; set; }
+        public double X { get; set; }
 
-        public int Y { get; set; }
+        public double Y { get; set; }
 
-        public int Z { get; set; }
+        public double Z { get; set; }
 
-        public int G0 { get; set; }
+        public double G0 { get; set; }
     }
 }
